[SofaQtQuickGUI] Move Manipulators file into a Manipulator directory (1/2)
