+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Manipulators/Manipulator2D_Rotation.h
@@ -17,10 +17,10 @@ You should have received a copy of the GNU General Public License
 along with sofaqtquick. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef MANIPULATOR3D_TRANSLATION_H
-#define MANIPULATOR3D_TRANSLATION_H
+#ifndef MANIPULATOR2D_ROTATION_H
+#define MANIPULATOR2D_ROTATION_H
 
-#include "SofaQtQuickGUI.h"
+#include <SofaQtQuickGUI/SofaQtQuickGUI.h>
 #include "Manipulator.h"
 
 #include <QObject>
@@ -31,24 +31,18 @@ namespace sofa
 namespace qtquick
 {
 
-/// \class An object or world space manipulator allowing us to translate things along X, Y and / or the Z axis
-class SOFA_SOFAQTQUICKGUI_API Manipulator3D_Translation : public Manipulator
+/// \class A view space manipulator allowing us to rotate things around the X or Y axis
+class SOFA_SOFAQTQUICKGUI_API Manipulator2D_Rotation : public Manipulator
 {
     Q_OBJECT
 
 public:
-    explicit Manipulator3D_Translation(QObject* parent = 0);
-    ~Manipulator3D_Translation();
+    explicit Manipulator2D_Rotation(QObject* parent = 0);
+    ~Manipulator2D_Rotation();
 
 public:
-    Q_PROPERTY(QString axis READ axis WRITE setAxis NOTIFY axisChanged)
-
-public:
-    QString axis() const {return myAxis;}
-    void setAxis(QString newAxis);
-
-signals:
-    void axisChanged(QString newAxis);
+    Q_INVOKABLE void setMark(float fromAngle, float toAngle);
+    Q_INVOKABLE void unsetMark();
 
 public slots:
     virtual void draw(const SofaViewer& viewer) const;
@@ -58,7 +52,9 @@ private:
     void internalDraw(const SofaViewer& viewer, bool isPicking = false) const;
 
 private:
-    QString myAxis;
+    bool  myDisplayMark;
+    float myFromMarkAngle;
+    float myToMarkAngle;
 
 };
 
@@ -66,4 +62,4 @@ private:
 
 }
 
-#endif // MANIPULATOR3D_TRANSLATION_H
+#endif // MANIPULATOR2D_ROTATION_H
