+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Rendering/ObjectRenderer.cpp
